odds_fetcher/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
httpx
apscheduler
python-dotenv
twilio
pandas 